agent/requirements.txt
websocket-client
pyshark
